#6863 - Fixes to make address sanitizer happy for internal runtime classes (#6880)
* Fixes to make address sanitizer happy.
Fixed initialization defects in StringStorage that could cause buffer overruns
Fixed memory leaks within RegionAllocator and BlockAllocator
Added system memory allocation tracking to all internal runtime tests.

* Clang Tidy / Format pass

* Fix formatting to use braces around if statements

+++ src/runtime/internal/memory_arena.h
@@ -128,10 +128,12 @@ void MemoryArena::initialize(void *user_context,
 }
 
 void MemoryArena::destroy(void *user_context) {
-    for (size_t i = blocks.size(); i--;) {
-        Block *block = lookup_block(user_context, i);
-        halide_abort_if_false(user_context, block != nullptr);
-        destroy_block(user_context, block);
+    if (!blocks.empty()) {
+        for (size_t i = blocks.size(); i--;) {
+            Block *block = lookup_block(user_context, i);
+            halide_abort_if_false(user_context, block != nullptr);
+            destroy_block(user_context, block);
+        }
     }
     blocks.destroy(user_context);
 }
